group routes and add layout to each
